src/requirements.txt
kopf[full-auth]~=1.36.0
kubernetes~=26.1.0
PyYAML~=6.0